+++ src/main/java/com/elmakers/mine/bukkit/action/builtin/SphereAction.java
@@ -2,7 +2,7 @@ package com.elmakers.mine.bukkit.action.builtin;
 
 public class SphereAction extends VolumeAction
 {
-	protected int innerRadiusSquared;
+	protected double innerRadiusSquared;
 
 	@Override
 	protected int getStartRadius() {
@@ -12,10 +12,10 @@ public class SphereAction extends VolumeAction
 
 	protected boolean containsPoint(int x, int y, int z)
 	{
-		float fx = (float)Math.abs(x) - 0.25f;
-		float fy = (float)Math.abs(y) - 0.25f;
-		float fz = (float)Math.abs(z) - 0.25f;
-		int distanceSquared = (int)((fx * fx) + (fy * fy) + (fz * fz));
+		double fx = (double)Math.abs(x) - 0.25;
+		double fy = (double)Math.abs(y) - 0.25;
+		double fz = (double)Math.abs(z) - 0.25;
+		double distanceSquared = ((fx * fx) + (fy * fy) + (fz * fz));
 		if (thickness > 0) {
 			return distanceSquared <= radiusSquared && distanceSquared >= innerRadiusSquared;
 		}
